Rename collisions.rs to collision.rs
